% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vadt.R
\name{vadt}
\alias{vadt}
\title{The visualizing Atlantis diagnostic tool}
\usage{
vadt(obj, anim = NULL)
}
\arguments{
\item{obj}{Object of class vat returned by create_vat}

\item{anim}{Directory to stored animated plot created by vat_animate function (defaults to NULL)}
}
\description{
The visualizing Atlantis diagnostic tool is an interactive Shiny-based tool for model tuning and calibration. It includes various plots, both aggregated and unaggregated, that are useful for diagnostic, tuning, and visualizing output. To use \code{vadt}, the user must first run the \code{create_vadt} function which will create an object of class \code{vadt} which can be fed to the \code{vadt} function. The user may also run the \code{animate_vadt} to create the animated GIFs (this is optional)
}
\examples{
\dontrun{
anim <- "/path/to/animinated/gifs"
obj <- create_vadt(outdir = "/atlantis/output_dir/",
fgfile = "/atlantis/functionalgroup.csv", 
ncout = "output_atlantis")
vadt(obj, anim = NULL)
}

}
\seealso{
\code{\link{create_vadt}}, \code{\link{animate_vadt}}
}

