% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vadt.R
\name{create_vadt}
\alias{create_vadt}
\title{Create a visualizing Atlantis diagnostic tool object}
\usage{
create_vadt(outdir, fgfile, biolprm, ncout, startyear, toutinc, diet = TRUE)
}
\arguments{
\item{outdir}{Path to Atlantis output directory}

\item{fgfile}{Functional group file}

\item{biolprm}{Biology parameter file}

\item{ncout}{Name of output ncdf4 file excluding nc suffix (i.e. name given after -o flag)}

\item{startyear}{Year that the model starts}

\item{toutinc}{Periodicity of writing output (in days)}

\item{diet}{Include diagnostic diet plots? default is TRUE}
}
\description{
This function creates a visualizing Atlantis diagnostic tool object which should be fed to the \code{vadt} function. This object can be quite large and this function may take a little while depending on how long you have run Atlantis for and the complexity of your model.
}
\examples{
\dontrun{
obj <- create_vadt(outdir = "out/", fgfile = "fg.csv", biolprm = "bio.prm", ncout = "atlout", startyear = 1948, toutinc = 30, diet = TRUE)
}

}
\seealso{
\code{\link{vadt}}, \code{\link{animate_vadt}}
}

