% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vat.R
\name{vat}
\alias{vat}
\title{The visualizing Atlantis toolbox}
\usage{
vat(obj, anim)
}
\arguments{
\item{obj}{Object of class vat returned by create_vat}

\item{anim}{Directory to stored animated plot created by vat_animate function}
}
\description{
The visualizing Atlantis toolbox is an interactive Shiny-based toolbox. It includes various plots, both aggregrated and unaggregated, that are useful for diagnostic, tuning, and visualizing output. To use \code{vat}, the user must first run the \code{create_vat} function which will create an object of class \code{vat} which can be fed to the \code{vat} function. The user must also run the \code{vat_animate} to create the animated GIFs required my the function. If you don't want to use the animated plots, you can just make the \code{anim} option and the warning message.
}
\seealso{
\code{\link{create_vat}}, \code{\link{animate_vat}}
anim <- "/path/to/animinated/gifs"
obj <- create_vat(outdir = "/atlantis/output_dir/",
fgfile = "/atlantis/functionalgroup.csv",
ncout = "output_atlantis")
vat(obj, anim)
}

